tle_api/requirements.txt
asttokens==3.0.0
boto3==1.36.2
botocore==1.36.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dsgp4==1.1.3
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
idna==3.10
ipykernel==6.29.5
ipynb==0.5.1
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
pyzmq==26.2.1
ratelimit==2.2.1
requests==2.32.3
s3transfer==0.11.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.44.0
